Move dom selectors to their own package
