+++ packages/dom/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@annotator/fragment-identifier",
+  "version": "0.0.0",
+  "description": "Utilities for annotation of the Document Object Model.",
+  "homepage": "https://annotator.apache.org",
+  "license": "Apache-2.0",
+  "author": "Apache Software Foundation",
+  "main": "index",
+  "repository": {
+    "type": "git",
+    "url": "https://gitbox.apache.org/repos/asf/incubator-annotator.git"
+  },
+  "devDependencies": {
+    "@annotator/selector": "0.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
